% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstackplot.R
\name{combine_plots}
\alias{combine_plots}
\title{Internal function to combine the plots and themes into one big plot}
\usage{
combine_plots(prepared_data)
}
\arguments{
\item{prepared_data}{the prepared data output from \code{\link[=prepare_data]{prepare_data()}}}
}
\value{
an object generated by \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}
}
\description{
Internal function to combine the plots and themes into one big plot
}
