% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstackplot.R
\name{combine_stackplot}
\alias{combine_stackplot}
\title{Internal function to combine the plots and themes into one big plot}
\usage{
combine_stackplot(prepared_stackplot)
}
\arguments{
\item{prepared_stackplot}{a nested data frame, typically the output from \code{\link[=prepare_stackplot]{prepare_stackplot()}}}
}
\value{
a plot object generated by \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}})
}
\description{
Internal function to combine the plots and themes into one big plot
}
