% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstackplot.R
\name{prepare_data}
\alias{prepare_data}
\title{Internal function to prepare the data for a ggstackplot}
\usage{
prepare_data(
  data,
  x,
  y,
  direction = c("guess", "horizontal", "vertical"),
  color = "black",
  overlap = 0,
  axis_size = 0.2,
  alternate_axes = TRUE
)
}
\arguments{
\item{data}{the data frame to plot}

\item{x}{the x variable(s) to plot, accepts \code{\link[dplyr:select]{dplyr::select()}} syntax}

\item{y}{the y variable(s) to plot, accepts \code{\link[dplyr:select]{dplyr::select()}} syntax}

\item{direction}{whether to make a horizontal or vertical ggstackplot (default is for the function to get based on the number of x and y variables provided)}

\item{color}{which color to make the plots}

\item{overlap}{fractional overlap with next plot (by default no overlap), if providing multiple values, should be 1 shorter than the number of stacked plots}

\item{axis_size}{what fraction of a plot to allocate for the fixed axis}

\item{alternate_axes}{whether to alternate the sides on which the stacked axes are plotted}
}
\value{
a nested data frame with each of stacked variables (.var), their plot configuration, and their data+
}
\description{
Internal function to prepare the data for a ggstackplot
}
