% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstackplot.R
\name{prepare_plots}
\alias{prepare_plots}
\title{Internal function to prepare plots to stack}
\usage{
prepare_plots(prepared_data, plot_template)
}
\arguments{
\item{prepared_data}{the prepared data output from \code{\link[=prepare_data]{prepare_data()}}}

\item{plot_template}{a template plot (ggplot object) to use for the stacked plots}
}
\value{
the prepared tibble with a plot column added
}
\description{
Internal function to prepare plots to stack
}
