% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstackplot.R
\name{prepare_stackplot}
\alias{prepare_stackplot}
\title{Prepare the stackplot}
\usage{
prepare_stackplot(
  data,
  x,
  y,
  direction = c("guess", "horizontal", "vertical"),
  color = "black",
  overlap = 0,
  axis_size = 0.2,
  alternate_axes = TRUE,
  switch_axes = TRUE,
  template = ggplot() + geom_line() + geom_point() + theme_bw()
)
}
\arguments{
\item{data}{the data frame to plot}

\item{x}{the x variable(s) to plot, accepts \code{\link[dplyr:select]{dplyr::select()}} syntax}

\item{y}{the y variable(s) to plot, accepts \code{\link[dplyr:select]{dplyr::select()}} syntax}

\item{direction}{whether to make a horizontal or vertical ggstackplot (default is for the function to get based on the number of x and y variables provided)}

\item{color}{which color to make the plots (also sets the plotwide color and fill aesthetics, overwrite in individual geoms in the \code{template} to overwrite this aesthetic)}

\item{overlap}{fractional overlap with next plot (by default no overlap), if providing multiple values, should be 1 shorter than the number of stacked plots}

\item{axis_size}{what fraction of a plot to allocate for the fixed axis}

\item{alternate_axes}{whether to alternate the sides on which the stacked axes are plotted}

\item{switch_axes}{whether to switch the stacked axis}

\item{template}{a template plot (ggplot object) to use for the stacked plots}
}
\value{
a nested data frame with all stacked variables (.var), their plot configuration, data, plot object, and theme object
}
\description{
This function is usually not called directly but can be used to assemble the parts of a stackplot first and then look at them or edit them individually before combining them with \code{\link[=combine_stackplot]{combine_stackplot()}}. The \code{\link[=ggstackplot]{ggstackplot()}} function just combines the two steps (\code{\link[=prepare_stackplot]{prepare_stackplot()}} and \code{\link[=combine_stackplot]{combine_stackplot()}}) into one step.
}
