% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstackplot.R
\name{prepare_themes}
\alias{prepare_themes}
\title{Internal function to prepare the themes for the stacked plots}
\usage{
prepare_themes(prepared_data)
}
\arguments{
\item{prepared_data}{the prepared data output from \code{\link[=prepare_data]{prepare_data()}}}
}
\value{
the prepared tibble with a theme column added
}
\description{
Internal function to prepare the themes for the stacked plots
}
