% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_ggplot.R
\name{stack_ggplot_vertically}
\alias{stack_ggplot_vertically}
\title{Stack a ggplot vertical}
\usage{
stack_ggplot_vertically()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
