% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstackplot-package.R
\docType{package}
\name{ggstackplot-package}
\alias{ggstackplot-package}
\title{ggstackplot: Create Overlapping Stacked Plots}
\description{
Easily create overlapping grammar of graphics plots for scientific data visualization. This style of plotting is particularly common in climatology and oceanography research communities.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Have you ever wanted to create (partly) overlapping line plots with matched color-coding of the data and axes? These kinds of plots are common in climatology and oceanography research but there is not an easy way to create them with ggplot facets. The ggstackplot package builds on \href{https://ggplot2.tidyverse.org/}{ggplot2} to provide a straightforward approach to building these kinds of plots while retaining the powerful grammar of graphics approach of ggplots. Check out the functionality provided by ggstackplots at \url{https://ggstackplot.kopflab.org}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ggstackplot.kopflab.org/}
  \item \url{https://github.com/kopflab/ggstackplot}
  \item Report bugs at \url{https://github.com/kopflab/ggstackplot/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu} (\href{https://orcid.org/0000-0002-2044-0201}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Tristan Caro \email{tristan.caro@colorado.edu} (\href{https://orcid.org/0000-0001-6177-7444}{ORCID})
  \item Jamie McFarlin \email{jamie.mcfarlin@uwyo.edu} (\href{https://orcid.org/0000-0002-8185-8185}{ORCID})
  \item Jon Raberg \email{Jonathan.Raberg@colorado.edu} (\href{https://orcid.org/0000-0003-2959-3631}{ORCID})
}

}
