% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstackplot-package.R, R/package.R
\docType{package}
\name{ggstackplot-package}
\alias{ggstackplot}
\alias{ggstackplot-package}
\title{ggstackplot: Create overlapping stacked ggplots}
\description{
This package makes it easy to create overlapping ggplots. This plotting style is particularly common in paleoclimatology and oceanography research communities.

This package makes it easy to create overlapping ggplots. This plotting style is particularly common in paleoclimatology and oceanography research communities.
}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu} (\href{https://orcid.org/0000-0002-2044-0201}{ORCID})

Authors:
\itemize{
  \item Tristan Caro \email{tristan.caro@colorado.edu}
  \item Jamie McFarlin \email{jamie.mcfarlin@uwyo.edu}
}

}
\keyword{internal}
