% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstackplot.R
\name{ggstackplot}
\alias{ggstackplot}
\title{Stack a ggplot}
\usage{
ggstackplot(
  data,
  x,
  y,
  direction = c("guess", "horizontal", "vertical"),
  color = "black",
  overlap = 0,
  axis_size = 0.2,
  alternate_axes = TRUE,
  switch_axes = FALSE,
  template = ggplot() + geom_line() + geom_point() + theme_bw() + theme(panel.grid =
    element_blank())
)
}
\arguments{
\item{data}{the data frame to plot}

\item{x}{the x variable(s) to plot, accepts \code{\link[dplyr:select]{dplyr::select()}} syntax}

\item{y}{the y variable(s) to plot, accepts \code{\link[dplyr:select]{dplyr::select()}} syntax}

\item{direction}{whether to make a horizontal or vertical ggstackplot (default is for the function to get based on the number of x and y variables provided)}

\item{color}{which color to make the plots (also sets the plotwide color and fill aesthetics, overwrite in individual geoms in the \code{template} to overwrite this aesthetic)}

\item{overlap}{fractional overlap with next plot (by default no overlap), if providing multiple values, should be 1 shorter than the number of stacked plots}

\item{axis_size}{what fraction of a plot to allocate for the fixed axis}

\item{alternate_axes}{whether to alternate the sides on which the stacked axes are plotted}

\item{switch_axes}{whether to switch the stacked axis}

\item{template}{a template plot (ggplot object) to use for the stacked plots}
}
\value{
a plot object generated by \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}})
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
